categories/ppc/ppc-medium-ocr/give/requirements.txt
Flask==3.1.0
Pillow==11.1.0
gunicorn==23.0.0
